Fixed some sign-compare warnings
Ticket: CFE-3415
Changelog: None

+++ libutils/hash_map_priv.h
@@ -53,7 +53,7 @@ typedef struct
 {
     HashMap *map;
     BucketListItem *cur;
-    int bucket;
+    size_t bucket;
 } HashMapIterator;
 
 HashMap *HashMapNew(MapHashFn hash_fn, MapKeyEqualFn equal_fn,
